move tests of JobScriptFiles to a dedicated file
